+++ extension/starter/src/main/java/org/camunda/bpm/spring/boot/starter/event/JobExecutorStartingEventListener.java
@@ -1,7 +1,7 @@
 package org.camunda.bpm.spring.boot.starter.event;
 
 import org.camunda.bpm.engine.impl.jobexecutor.JobExecutor;
-import org.camunda.bpm.spring.boot.starter.generic.SpringProcessEngineConfigurationTemplate;
+import org.camunda.bpm.spring.boot.starter.SpringProcessEngineConfigurationTemplate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
